Use position rather than ID to determine touch point
Using the id() of a TouchPoint does not give the same ID used to create
and move it. The ID of .press() and friends only determine the finger
or cursor to act on. This may have worked before due to an
implementation detail of the xcb or Mir QPAs.

See also https://doc.qt.io/qt-5/qtouchevent-touchpoint.html#id